appserver/requirements.txt
Django~=5.1.6
psycopg2-binary
django-cors-headers
djangorestframework